feat: integration tests organized by folders
