+++ tests/integration/multiple_servers/test.py
@@ -1,14 +1,14 @@
 import unittest
 import requests
-import _
+import integration
 
 BASE_URL_1 = "http://localhost:8080"
 BASE_URL_2 = "http://localhost:9000"
 
-response = _.get(BASE_URL_1)
+response = integration.get(BASE_URL_1)
 
 assert response.status_code == 200
 
-response = _.get(BASE_URL_2)
+response = integration.get(BASE_URL_2)
 
 assert response.status_code == 200
