+++ tests/integration/status_code/test.py
@@ -1,9 +1,9 @@
 import unittest
 import requests
-import _
+import integration
 
 BASE_URL = "http://localhost:8080"
 
-response = _.get(BASE_URL)
+response = integration.get(BASE_URL)
 
 assert response.status_code == 200
